test(pi-codex-conversion): cover scoped shutdown path
