+++ packages/pi-codex-conversion/tests/websocket-prewarm.test.ts
@@ -3,6 +3,8 @@ import assert from "node:assert/strict";
 import { buildCachedWebSocketRequestBody, buildRequestBody, prewarmOpenAICodexWebSocket } from "../src/providers/openai-codex-custom-provider.ts";
 import { acquireWebSocket, closeOpenAICodexWebSocketSessions } from "../src/providers/openai-codex/websocket.ts";
 import { createCodexTurnState } from "../src/providers/openai-codex/turn-state.ts";
+import { registerCodexEvents } from "../src/extension/events.ts";
+import { createCodexExtensionRuntime } from "../src/extension/runtime.ts";
 
 class FakeWebSocket {
 	static instances: FakeWebSocket[] = [];
@@ -113,11 +115,29 @@ test("WebSocket prewarm sends generate=false and seeds cached continuation", asy
 	}
 });
 
-test("closing one session leaves sibling cached WebSockets reusable", async () => {
+test("session shutdown leaves sibling cached WebSockets reusable", async () => {
 	const originalWebSocket = globalThis.WebSocket;
 	FakeWebSocket.instances = [];
 	(globalThis as typeof globalThis & { WebSocket: typeof WebSocket }).WebSocket = FakeWebSocket as never;
 	try {
+		const handlers = new Map<string, (...args: never[]) => unknown>();
+		const pi = {
+			on(event: string, handler: (...args: never[]) => unknown) {
+				handlers.set(event, handler);
+			},
+		};
+		const runtime = createCodexExtensionRuntime(pi as never);
+		registerCodexEvents(
+			pi as never,
+			runtime,
+			{} as never,
+			{ clearBackgroundWidget() {} } as never,
+			{ shutdown: async () => {} } as never,
+			{ shutdown() {} } as never,
+		);
+		const shutdown = handlers.get("session_shutdown");
+		assert.ok(shutdown);
+
 		const childA = await acquireWebSocket("wss://chatgpt.example/codex/responses", new Headers(), "child-a", undefined);
 		childA.release({ keep: true });
 		const childB = await acquireWebSocket("wss://chatgpt.example/codex/responses", new Headers(), "child-b", undefined);
@@ -125,7 +145,9 @@ test("closing one session leaves sibling cached WebSockets reusable", async () =
 
 		const socketA = FakeWebSocket.instances[0]!;
 		const socketB = FakeWebSocket.instances[1]!;
-		closeOpenAICodexWebSocketSessions("child-a");
+		await shutdown({ type: "session_shutdown", reason: "quit" } as never, {
+			sessionManager: { getSessionId: () => "child-a" },
+		} as never);
 
 		assert.deepEqual(socketA.closes, [{ code: 1000, reason: "session_shutdown" }]);
 		assert.deepEqual(socketB.closes, []);
@@ -134,7 +156,9 @@ test("closing one session leaves sibling cached WebSockets reusable", async () =
 		assert.equal(reusedB.reused, true);
 		reusedB.release({ keep: true });
 
-		closeOpenAICodexWebSocketSessions("child-b");
+		await shutdown({ type: "session_shutdown", reason: "quit" } as never, {
+			sessionManager: { getSessionId: () => "child-b" },
+		} as never);
 		assert.deepEqual(socketB.closes, [{ code: 1000, reason: "session_shutdown" }]);
 	} finally {
 		closeOpenAICodexWebSocketSessions();
